fix circular imports in operator
