+++ app/packages/operators/src/OperatorPrompt/index.tsx
@@ -2,7 +2,10 @@ import { createPortal } from "react-dom";
 import { useRecoilValue } from "recoil";
 import { showOperatorPromptSelector, useOperatorPrompt } from "../state";
 import { BaseStylesProvider } from "../styled-components";
-import { getPromptComponent, getPromptTarget } from "../utils";
+import { OperatorPromptType } from "../types";
+import OperatorModalPrompt from "./OperatorModalPrompt";
+import OperatorDrawerPrompt from "./OperatorDrawerPrompt";
+import { OPERATOR_PROMPT_AREAS } from "../constants";
 
 export default function OperatorPrompt() {
   const show = useRecoilValue(showOperatorPromptSelector);
@@ -25,4 +28,32 @@ function DynamicOperatorPrompt() {
   if (!prompt.resolvedIO.input && !prompt.resolvedIO.output) return null;
 
   return createPortal(<Component prompt={prompt} />, target);
+}
+
+const defaultTargetResolver = () => document.body;
+const targetResolverByName = {
+  DrawerView: (promptView: OperatorPromptType["promptView"]) => {
+    const promptArea =
+      promptView.placement === "left"
+        ? OPERATOR_PROMPT_AREAS.DRAWER_LEFT
+        : OPERATOR_PROMPT_AREAS.DRAWER_RIGHT;
+    return document.getElementById(promptArea);
+  },
+};
+export function getPromptTarget(operatorPrompt: OperatorPromptType) {
+  const { promptView } = operatorPrompt;
+  const targetResolver =
+    targetResolverByName[promptView?.name] || defaultTargetResolver;
+  return targetResolver(promptView);
+}
+
+const defaultPromptComponentResolver = () => OperatorModalPrompt;
+const promptComponentByName = {
+  DrawerView: () => OperatorDrawerPrompt,
+};
+export function getPromptComponent(operatorPrompt: OperatorPromptType) {
+  const { promptView } = operatorPrompt;
+  const targetResolver =
+    promptComponentByName[promptView?.name] || defaultPromptComponentResolver;
+  return targetResolver(promptView);
 }
